[UPDATE] rename mypage -> mainpage
